initial commit with refactoring updates
