+++ event/mock/handler.go
@@ -10,25 +10,29 @@ import (
 	"sync"
 )
 
-// Ensure, that HandlerMock does implement event.Handler.
+var (
+	lockHandlerMockHandle sync.RWMutex
+)
+
+// Ensure, that HandlerMock does implement Handler.
 // If this is not the case, regenerate this file with moq.
 var _ event.Handler = &HandlerMock{}
 
 // HandlerMock is a mock implementation of event.Handler.
 //
-// 	func TestSomethingThatUsesHandler(t *testing.T) {
+//     func TestSomethingThatUsesHandler(t *testing.T) {
 //
-// 		// make and configure a mocked event.Handler
-// 		mockedHandler := &HandlerMock{
-// 			HandleFunc: func(ctx context.Context, reindexRequested *models.ReindexRequested) error {
-// 				panic("mock out the Handle method")
-// 			},
-// 		}
+//         // make and configure a mocked event.Handler
+//         mockedHandler := &HandlerMock{
+//             HandleFunc: func(ctx context.Context, reindexRequested *models.ReindexRequested) error {
+// 	               panic("mock out the Handle method")
+//             },
+//         }
 //
-// 		// use mockedHandler in code that requires event.Handler
-// 		// and then make assertions.
+//         // use mockedHandler in code that requires event.Handler
+//         // and then make assertions.
 //
-// 	}
+//     }
 type HandlerMock struct {
 	// HandleFunc mocks the Handle method.
 	HandleFunc func(ctx context.Context, reindexRequested *models.ReindexRequested) error
@@ -43,7 +47,6 @@ type HandlerMock struct {
 			ReindexRequested *models.ReindexRequested
 		}
 	}
-	lockHandle sync.RWMutex
 }
 
 // Handle calls HandleFunc.
@@ -58,9 +61,9 @@ func (mock *HandlerMock) Handle(ctx context.Context, reindexRequested *models.Re
 		Ctx:              ctx,
 		ReindexRequested: reindexRequested,
 	}
-	mock.lockHandle.Lock()
+	lockHandlerMockHandle.Lock()
 	mock.calls.Handle = append(mock.calls.Handle, callInfo)
-	mock.lockHandle.Unlock()
+	lockHandlerMockHandle.Unlock()
 	return mock.HandleFunc(ctx, reindexRequested)
 }
 
@@ -75,8 +78,8 @@ func (mock *HandlerMock) HandleCalls() []struct {
 		Ctx              context.Context
 		ReindexRequested *models.ReindexRequested
 	}
-	mock.lockHandle.RLock()
+	lockHandlerMockHandle.RLock()
 	calls = mock.calls.Handle
-	mock.lockHandle.RUnlock()
+	lockHandlerMockHandle.RUnlock()
 	return calls
 }
